requirements.txt
Flask==3.0.0
gunicorn==21.2.0
openai==0.27.8
firebase-admin==6.2.0
inotify==0.2.10
bcrypt==4.0.1
pytest==7.4.2
flask-pytest==0.0.5
flask-cors==4.0.0
python-dotenv==1.0.0